CI: Fix pre-commit step and split tox
Pre-commit action is actually 3.0.0. Also because we run black through
pre-commit we do not need to run it wia Tox. We split tox for smaller
components. This is also nice as then we notice fails by looking steps
rather that going check output. For this to work we need to continue on
error.