Add first home view with templates
Add basic template structure which we can use. Also add first view which
we can start to use.

I would like to use also some html formatter, but only found prettier
and I did not like that it is node app. Maybe it is right choice, but
let's think that later. Just using indent_size 2 for now for html files.